Move shares db objects to shares_base
