+++ backend/dataall/modules/s3_datasets_shares/services/dataset_sharing_service.py
@@ -4,7 +4,7 @@ from dataall.core.environment.services.environment_service import EnvironmentSer
 from dataall.base.context import get_context
 from dataall.base.db import exceptions
 from dataall.base.aws.sts import SessionHelper
-from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObject
+from dataall.modules.shares_base.db.share_object_models import ShareObject
 from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
 from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
 from dataall.modules.shares_base.services.share_permissions import SHARE_OBJECT_APPROVER
