+++ backend/dataall/modules/s3_datasets_shares/services/share_managers/lf_share_manager.py
@@ -21,7 +21,7 @@ from dataall.modules.shares_base.services.shares_enums import (
 )
 from dataall.modules.s3_datasets.db.dataset_models import DatasetTable, S3Dataset
 from dataall.modules.s3_datasets_shares.services.dataset_sharing_alarm_service import DatasetSharingAlarmService
-from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObjectItem, ShareObject
+from dataall.modules.shares_base.db.share_object_models import ShareObjectItem, ShareObject
 from dataall.modules.s3_datasets_shares.services.share_managers.share_manager_utils import ShareErrorFormatter
 
 logger = logging.getLogger(__name__)
