+++ backend/dataall/modules/s3_datasets_shares/services/share_processors/lakeformation_process_share.py
@@ -14,7 +14,7 @@ from dataall.modules.s3_datasets_shares.services.share_managers import LFShareMa
 from dataall.modules.s3_datasets_shares.aws.ram_client import RamClient
 from dataall.modules.s3_datasets_shares.services.share_object_service import ShareObjectService
 from dataall.modules.s3_datasets.db.dataset_models import DatasetTable, S3Dataset
-from dataall.modules.s3_datasets_shares.db.share_object_models import ShareObject
+from dataall.modules.shares_base.db.share_object_models import ShareObject
 from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
 from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
 from dataall.modules.s3_datasets_shares.services.share_managers.share_manager_utils import ShareErrorFormatter
