+++ backend/dataall/modules/shares_base/db/share_object_models.py
@@ -52,9 +52,9 @@ class ShareObjectItem(Base):
     updated = Column(DateTime, nullable=True, onupdate=datetime.now)
     deleted = Column(DateTime, nullable=True)
     owner = Column(String, nullable=False)
-    GlueDatabaseName = Column(String, nullable=True)
-    GlueTableName = Column(String, nullable=True)
-    S3AccessPointName = Column(String, nullable=True)
+    GlueDatabaseName = Column(String, nullable=True)  # TODO: remove
+    GlueTableName = Column(String, nullable=True)  # TODO: remove
+    S3AccessPointName = Column(String, nullable=True)  # TODO: remove
     status = Column(String, nullable=False, default=ShareItemStatus.PendingApproval.value)
     action = Column(String, nullable=True)
     healthStatus = Column(String, nullable=True)
